Add Item callbacks for unbind and isRecyclable
In order to accomplish a callback for unbinding the viewholder, the
viewholder must hold a reference to the Item. I was nervous about
coupling them earlier, but I think it's the only way that makes sense.

I also changed GroupAdapter to specify com.genius.groupie.ViewHolder,
which removes the need for some casting, and improved the javadoc a little.