Annotate everything with NonNull / Nullable for Kotlin
In order to guarantee non-nullable on more methods, the behavior of some methods has changed.
In particular, a few methods like getItem() / getGroup() which used to return null if not
present now throw IndexOutOfBoundsException. You can still check if an Item is present by
checking for its index being > -1.
Additionally, setGroupDataObserver is now register/unregisterGroupDataObserver and its arguments
are @NonNull.