Use experimental ViewHolder support (#125)
* Use experimental ViewHolder support

This would replace the current (broken) way of using Kotlin Android Extensions. In the current way of doing things, the view lookups (findViewById) aren't cached in ViewHolders, which is bad for performance.

Viewholder view lookups are (incidentally) also nicer now.  Instead of viewholder.itemView.title, you now have just viewholder.title.

In order not to make the core library depend on Kotlin, there will now be an additional library:
groupie-kotlin-android-extensions

Unlike the data binding dependency, this is an add on.  You will need *both* the core groupie library and groupie-kotlin-android-extensions.  In order to stay consistent, we could also change the data binding library to an add on. It would be weird, but there is theoretically no reason not to use both.

* Update novoda-bintray-release, otherwise we get build errors

* Switch to Kotlin 1.1.6-EAP to allow use of LayoutContainer as interface

* Perhaps circle ci would like this?

* Or perhaps circle will like this