[WIP] async update
Using AsyncListDiffer won't work here as long as we extend adapter
and accept other entry points for updates.  Will have to create a new
object (OneWayGroupie or something) which translates between groups
and items, then the diff can communicate with a normal ListAdapter,
or any adapter.

In this case, updates to sections and individual items won't work. Users
will have to completely commit to a one-way data flow.